refactor index.js into one-component-per-file
